refactor: organização e padronização de nomes e caminhos e devidas adaptações
