+++ src/main/resources/static/JS/imovel/gerenciarImovel.js
