+++ src/main/resources/templates/dashboard/dashboard.html
@@ -8,7 +8,7 @@
     <!-- Material Icons -->
 	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-    <link rel="stylesheet" href="css/dashboard.css">
+    <link rel="stylesheet" href="css/dashboard/dashboard.css">
 </head>
 <body>
 
@@ -114,7 +114,7 @@
 
 <div id="toastSucesso" class="toast-sucesso"></div>
 
-<script src="/JS/dashboard.js"></script>
+<script src="/JS/dashboard/dashboard.js"></script>
 
 </body>
