+++ src/main/resources/templates/imovel/cadastroImovel.html
@@ -8,7 +8,7 @@
 	<!-- Material Icons -->
 	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-    <link rel="stylesheet" href="/css/cadastroImovel.css">
+    <link rel="stylesheet" href="/css/imovel/cadastroImovel.css">
 </head>
 <body>
 
@@ -212,6 +212,6 @@
     </footer>
 </main>
 <script src="https://unpkg.com/imask"></script>
-<script src="/js/cadastroImovel.js"></script>
+<script src="/js/imovel/cadastroImovel.js"></script>
 </body>
 </html>
