+++ src/main/resources/templates/imovel/editarImovel.html
@@ -6,7 +6,7 @@
     <title>Cadastrar Novo Imóvel - InvistaIX</title>
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
 	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
-    <link rel="stylesheet" href="/css/editarImovel.css">
+    <link rel="stylesheet" href="/css/imovel/editarImovel.css">
 </head>
 <body>
 	
@@ -241,6 +241,6 @@
 
 </div>
 <script src="https://unpkg.com/imask"></script>
-<script src="/js/editarImovel.js"></script>
+<script src="/js/imovel/editarImovel.js"></script>
 </body>
 </html>
