+++ src/main/resources/templates/imovel/gerenciarImovel.html
@@ -682,6 +682,6 @@
 <div id="toastSucesso" class="toast-sucesso"></div>
 <script src="https://unpkg.com/imask@7.6.0/dist/imask.min.js"></script>
 <script src="https://cdn.jsdelivr.net/npm/vanillajs-datepicker@1.3.4/dist/js/datepicker-full.min.js"></script>
-<script src="/JS/gerenciarImovel.js"></script>
+<script src="/JS/imovel/gerenciarImovel.js"></script>
 </body>
 </html>
