+++ src/main/resources/templates/imovel/graficosImovel.html
@@ -8,7 +8,7 @@
   	<link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
   	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
   	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/vanillajs-datepicker@1.3.4/dist/css/datepicker.min.css">
-  	<link rel="stylesheet" href="/css/graficos.css">
+  	<link rel="stylesheet" href="/css/imovel/graficosImovel.css">
 </head>
 
 <body>
@@ -102,7 +102,7 @@
 
 <script src="https://code.highcharts.com/highcharts.js"></script>
 <script src="https://code.highcharts.com/modules/accessibility.js"></script>
-<script src="/JS/graficos.js"></script>
+<script src="/JS/imovel/graficosImovel.js"></script>
 
 </body>
 </html>
