+++ src/main/resources/templates/usuario/cadastroUsuario.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
     <title>Cadastro - invistaIX</title>
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;600;800&display=swap" rel="stylesheet"/>
-	<link rel="stylesheet" href="css/cadastroUsuario.css">
+	<link rel="stylesheet" href="css/usuario/cadastroUsuario.css">
 </head>
 <body>
 <header>
@@ -69,6 +69,6 @@
 <footer>
     © 2025 InvistaIX. Todos os direitos reservados.
 </footer>
-<script src="/JS/cadastro.js"></script>
+<script src="/JS/usuario/cadastroUsuario.js"></script>
 </body>
 </html>
