+++ src/main/resources/templates/usuario/perfilUsuario.html
@@ -8,7 +8,7 @@
   <!-- Material Icons -->
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
   <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-  <link rel="stylesheet" href="css/perfilusuario.css">
+  <link rel="stylesheet" href="css/usuario/perfilUsuario.css">
 </head>
 <body>
